Ensure all forward pass operations use float32 throughout
Added .astype(np.float32) to all intermediate computations to ensure the entire model uses float32:
- embeddings.py: Ensure forward pass output is float32
- feed_forward.py: Cast all intermediate operations to float32
- single_head_attention.py: Cast Q, K, V, attention scores, and outputs to float32
- transformer_block.py: Cast layer normalization output to float32
- output_layer.py: Cast logits and softmax outputs to float32

This ensures consistent float32 usage throughout the entire forward and backward passes for optimal performance.

+++ src/embeddings/embeddings.py
@@ -86,10 +86,10 @@ class EmbeddingLayer:
         token_embeddings = self.embeddings[padded_token_ids]  # shape (batch_size, max_len, embedding_dim)
         token_embeddings = token_embeddings * np.sqrt(self.embedding_dim)  # scale embeddings
 
-        pos_enc = self.positional_encodings[:self.max_len, :]  # shape (max_len, embedding_dim)
+        pos_enc = self.positional_encodings[:self.max_len, :].astype(np.float32)  # shape (max_len, embedding_dim)
 
         # Add positional encoding to each sequence
-        output = token_embeddings + pos_enc[np.newaxis, :, :]
+        output = (token_embeddings + pos_enc[np.newaxis, :, :]).astype(np.float32)
 
         return output  # shape (batch_size, max_len, embedding_dim)
